Implement basic Dashiki extension
This extension configures the sub-namespace Dashiki in the Config
namespace.  We would like to deploy this on meta so we can create more
valid Dashiki configuration pages as Config:Dashiki:NewDashboardConfig
instead of the current way we're just faking the Config: prefix.  As a
bonus of extending JsonConfig's View, we get nicer display for our
configs and we can start to add useful shortcuts for Dashiki dashboard
maintainers.

Bug: T125403
Change-Id: Ic21c91f43e06f9261d7636ea554702c7242377ee